requirements.txt
flask>=3.0.0
pillow>=10.0.0
torch>=2.0.0
git+https://github.com/openai/CLIP.git
requests>=2.31.0
playwright>=1.40.0
numpy>=1.24.0
